prettify and lint solidity code

+++ sc/contracts/interfaces/ISafeStorage.sol
@@ -4,8 +4,9 @@ pragma solidity >=0.7.6;
 pragma abicoder v2;
 
 interface ISafeStorage {
-    function execute(address _receipts, uint _value, bytes memory _data)
-        external
-        payable
-        returns (bool success, bytes memory result);
+    function execute(
+        address _receipts,
+        uint256 _value,
+        bytes memory _data
+    ) external payable returns (bool success, bytes memory result);
 }
