+++ sc/contracts/multisig/Multisig.sol
@@ -7,11 +7,10 @@ import "../interfaces/ITimelock.sol";
 import "../libs/TimelockLibrary.sol";
 
 contract Multisig is Signable {
-
     enum Status {
         EMPTY, // zero state
         INITIALIZED, // created with one sign
-        CANCELLED,  // canceled by consensus
+        CANCELLED, // canceled by consensus
         QUEUED, // approved and send to timelock
         EXECUTED // executed
     }
@@ -19,42 +18,40 @@ contract Multisig is Signable {
     struct Proposal {
         // @dev actual signs
         uint256 signs;
-
         Status status;
         /// @notice Creator of the proposal
         address proposer;
         /// @notice The timestamp that the proposal will be available for execution, set once the vote succeeds
-        uint eta;
+        uint256 eta;
         /// @notice the ordered list of target addresses for calls to be made
         address[] targets;
         /// @notice The ordered list of values (i.e. msg.value) to be passed to the calls to be made
-        uint[] values;
+        uint256[] values;
         /// @notice The ordered list of function signatures to be called
         string[] signatures;
         /// @notice The ordered list of calldata to be passed to each call
         bytes[] calldatas;
-
         address callFrom;
-
         string description;
-
         uint256 initiatedAt;
     }
 
-    mapping (uint => Proposal) public proposals;
-    mapping (address => mapping (uint => bool)) public votedBy;
+    mapping(uint256 => Proposal) public proposals;
+    mapping(address => mapping(uint256 => bool)) public votedBy;
 
     /// @notice The total number of proposals
-    uint public proposalCount;
+    uint256 public proposalCount;
 
     address public timelock;
 
-    event ProposalInitialized(uint id, address proposer);
-    event Signed(uint id, address signer);
-    event Executed(uint id);
-    event Cancelled(uint id);
+    event ProposalInitialized(uint256 id, address proposer);
+    event Signed(uint256 id, address signer);
+    event Executed(uint256 id);
+    event Cancelled(uint256 id);
 
-    constructor(address _timelock, address[] memory _accounts) Signable(_accounts) {
+    constructor(address _timelock, address[] memory _accounts)
+        Signable(_accounts)
+    {
         require(_timelock != address(0), "Timelock zero");
 
         timelock = _timelock;
@@ -62,7 +59,7 @@ contract Multisig is Signable {
 
     function createAndSign(
         address[] memory targets,
-        uint[] memory values,
+        uint256[] memory values,
         string[] memory signatures,
         bytes[] memory calldatas,
         string memory description,
@@ -81,18 +78,15 @@ contract Multisig is Signable {
 
         proposalCount++;
 
-        uint proposalId = proposalCount;
+        uint256 proposalId = proposalCount;
         proposals[proposalId] = proposal;
 
         emit ProposalInitialized(proposalId, msg.sender);
         emit Signed(proposalId, msg.sender);
     }
 
-    function sign(uint _proposalId) external onlySigner {
-        require(
-             getStatus(_proposalId) == Status.INITIALIZED,
-            "Wrong status"
-        );
+    function sign(uint256 _proposalId) external onlySigner {
+        require(getStatus(_proposalId) == Status.INITIALIZED, "Wrong status");
         require(!votedBy[msg.sender][_proposalId], "Already signed");
 
         votedBy[msg.sender][_proposalId] = true;
@@ -103,13 +97,23 @@ contract Multisig is Signable {
             proposal.status = Status.QUEUED; // block status
             proposal.eta = ITimelock(timelock).delay() + block.timestamp;
             TimelockLibrary.Transaction memory txn;
-            for (uint i; i < proposal.targets.length; i++) {
+            for (uint256 i; i < proposal.targets.length; i++) {
                 txn.target = proposal.targets[i];
                 txn.value = proposal.values[i];
                 txn.signature = proposal.signatures[i];
                 txn.data = proposal.calldatas[i];
                 txn.eta = proposal.eta;
-                txn.hash = keccak256(abi.encode(_proposalId, i, txn.target, txn.value, txn.signature, txn.data, txn.eta));
+                txn.hash = keccak256(
+                    abi.encode(
+                        _proposalId,
+                        i,
+                        txn.target,
+                        txn.value,
+                        txn.signature,
+                        txn.data,
+                        txn.eta
+                    )
+                );
                 txn.callFrom = proposal.callFrom;
 
                 ITimelock(timelock).queueTransaction(txn);
@@ -120,40 +124,52 @@ contract Multisig is Signable {
     }
 
     // _paidFromStorage - if call withdraw or ether should be paid from storage contract
-    function execute(uint _proposalId, bool _paidFromStorage) public payable onlySigner {
+    function execute(uint256 _proposalId, bool _paidFromStorage)
+        public
+        payable
+        onlySigner
+    {
         if (_paidFromStorage) {
             require(msg.value == 0, "Pay from storage");
         }
 
-        require(
-            getStatus(_proposalId) == Status.QUEUED,
-            "Wrong status"
-        );
+        require(getStatus(_proposalId) == Status.QUEUED, "Wrong status");
 
         Proposal storage proposal = proposals[_proposalId];
         proposal.status = Status.EXECUTED; // block status
         TimelockLibrary.Transaction memory txn;
-        for (uint i; i < proposal.targets.length; i++) {
+        for (uint256 i; i < proposal.targets.length; i++) {
             txn.target = proposal.targets[i];
             txn.value = proposal.values[i];
             txn.signature = proposal.signatures[i];
             txn.data = proposal.calldatas[i];
             txn.eta = proposal.eta;
-            txn.hash = keccak256(abi.encode(_proposalId, i, txn.target, txn.value, txn.signature, txn.data, txn.eta));
+            txn.hash = keccak256(
+                abi.encode(
+                    _proposalId,
+                    i,
+                    txn.target,
+                    txn.value,
+                    txn.signature,
+                    txn.data,
+                    txn.eta
+                )
+            );
             txn.callFrom = proposal.callFrom;
 
-            ITimelock(timelock).executeTransaction{value: (_paidFromStorage) ? 0 : txn.value}(txn);
+            ITimelock(timelock).executeTransaction{
+                value: (_paidFromStorage) ? 0 : txn.value
+            }(txn);
         }
 
         emit Executed(_proposalId);
     }
 
-    function cancel(uint _proposalId) external onlySigner {
+    function cancel(uint256 _proposalId) external onlySigner {
         Status status = getStatus(_proposalId);
 
         require(
-            status == Status.INITIALIZED ||
-            status == Status.QUEUED,
+            status == Status.INITIALIZED || status == Status.QUEUED,
             "Wrong status"
         );
 
@@ -161,13 +177,23 @@ contract Multisig is Signable {
         proposal.status = Status.CANCELLED;
 
         TimelockLibrary.Transaction memory txn;
-        for (uint i; i < proposal.targets.length; i++) {
+        for (uint256 i; i < proposal.targets.length; i++) {
             txn.target = proposal.targets[i];
             txn.value = proposal.values[i];
             txn.signature = proposal.signatures[i];
             txn.data = proposal.calldatas[i];
             txn.eta = proposal.eta;
-            txn.hash = keccak256(abi.encode(_proposalId, i, txn.target, txn.value, txn.signature, txn.data, txn.eta));
+            txn.hash = keccak256(
+                abi.encode(
+                    _proposalId,
+                    i,
+                    txn.target,
+                    txn.value,
+                    txn.signature,
+                    txn.data,
+                    txn.eta
+                )
+            );
             txn.callFrom = proposal.callFrom;
 
             ITimelock(timelock).cancelTransaction(txn);
@@ -176,7 +202,7 @@ contract Multisig is Signable {
         emit Cancelled(_proposalId);
     }
 
-    function getActions(uint _proposalId)
+    function getActions(uint256 _proposalId)
         external
         view
         returns (
@@ -190,7 +216,7 @@ contract Multisig is Signable {
         return (p.targets, p.values, p.signatures, p.calldatas);
     }
 
-    function getStatus(uint _proposalId) public view returns (Status) {
+    function getStatus(uint256 _proposalId) public view returns (Status) {
         Proposal memory p = proposals[_proposalId];
 
         if (p.status == Status.CANCELLED) {
